Refactor state machine tests, restructure test folder
