+++ src/isar/config/default.ini
@@ -69,7 +69,7 @@ url = https://echohubapi.equinor.com/api
 [logging]
 azure_storage_logging_level = WARNING
 transitions_core_logging_level = WARNING
-state_transitions_log_length = 5
+state_transitions_log_length = 20
 
 [maps]
 eq_robot_default_map_name = default_map
