+++ tests/isar/services/readers/test_map_reader.py
