+++ tests/isar/state_machine/states/test_idle.py
@@ -3,7 +3,7 @@ import pytest
 from isar.models.mission import Mission
 from isar.state_machine.state_machine import StateMachine, States
 from robot_interface.models.mission import DriveToPose
-from tests.test_utilities.mock_models.mock_robot_variables import mock_pose
+from tests.mocks.robot_variables import mock_pose
 
 
 @pytest.mark.parametrize(
